libzbd: Add capacity support
Upcoming NVMe ZNS zoned devices will define a zone "capacity" which
can be smaller than the zone size and indicates the amount of sectors
that are usable in the zone.

To accomodate this once kernel support is added for these new devices,
add the capacity field to the zbd_zone data structure.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>