lib: fix BLKGETNRZONES use compilation errors
Fix compilation errors with kernel versions preceding 4.20 that do not
define the BLKGETNRZONES ioctl command.
Introduce the helper function zbd_get_nr_zones() to obtain the number
of zones of a device. This helper is conditionally defined depending
on the existence of the BLKGETNRZONES ioctl call. For kernel versions
not defining this ioctl, the number of zones of the device is calculated
using the device capacity and zone size.

This helper is used in zbd_do_get_info() to initialize the device
information nr_zones field.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>