tools: zbd: Introduce dump command
Introduce the new dump command. This new command saves a drive zone
information and zone data to files. For a zoned block device /dev/XXX,
the zone information is saved in the file XXX_zone_info.dump and the
zone data is saved in the file XXX_zone_data.dump.

The zone information file format is:
- A 128 B struct zbd_info header
- An array of 64B struct zbd_zone data structures, one for each zone of
  the device.
The report command can be used to inspect a zone information file
similarly to doing a report on a zoned device.

The size of the zone data file is always equal to the device capacity.
The data copied are all valid data sectors of the device within the
offset and offset+length range defined by the command line (-ofst and
-len options). By default, all valid data sectors are copied, that is,
all sectors of conventional zones and all written sectors of sequential
zones.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>