build: fix library versioning
Building the library using the -release option with the package version
as the library version results in a non binary compatible library,
forcing users to recompile their application against newer versions of
libzbd, even if no interface or ABI change is introduced.

Fix this by using the -version-number libtool option, which ensures
that applications compiled against previous versions of libzbd will
automatically load the latest version of libzbd without any application
recompile needed. This is, of course, limited to compatibility with
versions that share the same major version number.

With this change, the installed library .so file name changes from
libzbd-1.x.y.so to libzbd.so.1.x.y. The symbolic link libzbd.so also now
points to this new file.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>