lib: improve zbd_zones_operation()
The BLKOPENZONE, BLKCLOSEZONE and BLKFINISHZONE ioctl commands were
introduced in kernel 5.5. If they are not defined on the system used to
compile libzbd, manually define these operations to generate code
portable to system with newer kernels supporting these operations.

While at it, fix error code return to be consistent: errno is set and -1
is returned to indicate the operation failure.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>